+++ lib/service/lint.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/Jeffail/benthos/v3/internal/docs"
 	"github.com/Jeffail/benthos/v3/lib/config"
 	"github.com/fatih/color"
 	"github.com/urfave/cli/v2"
@@ -39,9 +40,9 @@ type pathLint struct {
 	err    string
 }
 
-func lintFile(path string) (pathLints []pathLint) {
+func lintFile(path string, rejectDeprecated bool) (pathLints []pathLint) {
 	conf := config.New()
-	lints, err := config.Read(path, true, &conf)
+	lints, err := config.ReadV2(path, true, rejectDeprecated, &conf)
 	if err != nil {
 		pathLints = append(pathLints, pathLint{
 			source: path,
@@ -58,7 +59,7 @@ func lintFile(path string) (pathLints []pathLint) {
 	return
 }
 
-func lintMDSnippets(path string) (pathLints []pathLint) {
+func lintMDSnippets(path string, rejectDeprecated bool) (pathLints []pathLint) {
 	rawBytes, err := os.ReadFile(path)
 	if err != nil {
 		pathLints = append(pathLints, pathLint{
@@ -97,7 +98,9 @@ func lintMDSnippets(path string) (pathLints []pathLint) {
 				err:    err.Error(),
 			})
 		} else {
-			lints, err := config.Lint(configBytes, conf)
+			lintCtx := docs.NewLintContext()
+			lintCtx.RejectDeprecated = rejectDeprecated
+			lints, err := config.LintV2(lintCtx, configBytes)
 			if err != nil {
 				pathLints = append(pathLints, pathLint{
 					source: path,
@@ -135,10 +138,18 @@ Exits with a status code 1 if any linting errors are detected:
 
 If a path ends with '...' then Benthos will walk the target and lint any
 files with the .yaml or .yml extension.`[1:],
+		Flags: []cli.Flag{
+			&cli.BoolFlag{
+				Name:  "deprecated",
+				Value: false,
+				Usage: "Print linting errors for the presence of deprecated fields.",
+			},
+		},
 		Action: func(c *cli.Context) error {
 			var targets []string
 			for _, p := range c.Args().Slice() {
 				var recurse bool
+				// TODO: V4 support wildcards
 				if p, recurse = resolveLintPath(p); recurse {
 					if err := filepath.Walk(p, func(path string, info os.FileInfo, werr error) error {
 						if werr != nil {
@@ -164,6 +175,8 @@ files with the .yaml or .yml extension.`[1:],
 				targets = append(targets, conf)
 			}
 
+			rejectDeprecated := c.Bool("deprecated")
+
 			var pathLintMut sync.Mutex
 			var pathLints []pathLint
 			threads := runtime.NumCPU()
@@ -181,9 +194,9 @@ files with the .yaml or .yml extension.`[1:],
 						}
 						var lints []pathLint
 						if path.Ext(target) == ".md" {
-							lints = lintMDSnippets(target)
+							lints = lintMDSnippets(target, rejectDeprecated)
 						} else {
-							lints = lintFile(target)
+							lints = lintFile(target, rejectDeprecated)
 						}
 						if len(lints) > 0 {
 							pathLintMut.Lock()